third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/read-updates-value.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Succesful read should update characteristic.value';
const EXPECTED_VALUE = [0, 1, 2];
let characteristic, fake_characteristic;
promise_test(() => getMeasurementIntervalCharacteristic()
  .then(_ => ({characteristic, fake_characteristic} = _))
  .then(() => assert_equals(characteristic.value, null))
  .then(() =>
      fake_characteristic.setNextReadResponse(GATT_SUCCESS, EXPECTED_VALUE))
  .then(() => characteristic.readValue())
  .then(() => assert_array_equals(
       new Uint8Array(characteristic.value.buffer), EXPECTED_VALUE))
  .then(() => navigator.bluetooth.test.allResponsesConsumed())
  .then(consumed => assert_true(consumed)), test_desc);
</script>
